refactor: separate study React hooks from state
