+++ src/features/deck/containers/DeckListContainer.tsx
@@ -6,7 +6,7 @@ import * as selector from "@src/selector";
 import { useActions } from "@src/shared/hooks/useActions";
 import { DeckListTemplate } from "@src/features/deck/components/templates/DeckListTemplate";
 import { getLegacyStudyCandidate } from "@src/features/study/hooks/useLegacyStudySession";
-import { useStudyStore } from "@src/features/study/state/studyStore";
+import { useStudyStore } from "@src/features/study/hooks/useStudyStore";
 
 export const DeckListContainer: React.FC = () => {
   const actions = useActions();
