+++ src/features/study/hooks/useLegacyStudySession.ts
@@ -3,7 +3,8 @@ import React from "react";
 import { useDispatch } from "react-redux";
 
 import * as type from "@src/action/type";
-import { type LegacyStudyCandidate, type LegacyStudyFields, useStudyStore } from "@src/features/study/state/studyStore";
+import { useStudyStore } from "@src/features/study/hooks/useStudyStore";
+import { type LegacyStudyCandidate, type LegacyStudyFields } from "@src/features/study/state/studyStore";
 
 export const getLegacyStudyCandidate = (deck: Deck): LegacyStudyCandidate | undefined => {
   const legacyDeck = deck as Deck & Partial<LegacyStudyFields>;
